requirements.txt
appdirs==1.4.4
attrs==22.1.0
blinker==1.4
boto3~=1.18.13
botocore~=1.21.13
brotlipy==0.7.0
bcrypt~=4.0.1
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
docxcompose==1.3.6
docxtpl==0.16.4
dynamorm==0.11.0
Flask~=2.3.2
flask-lambda==0.0.4
idna==3.3
importlib-metadata==5.0.0
itsdangerous==2.1.0
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==4.16.0
lxml==4.9.1
MarkupSafe==2.1.0
marshmallow==3.15.0
oauthlib==3.2.2
packaging==21.3
phonenumbers==8.13.8
pycountry==22.3.5
pycparser==2.21
pyee==8.2.2
PyJWT==2.3.0
PyYAML==5.4.1
pyparsing==3.0.7
pyrsistent==0.19.1
python-baseconv==1.2.2
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.5.2
six==1.16.0
svix-ksuid==0.6.0
tqdm==4.64.1
twilio==7.7.0
urllib3==1.26.8
websockets==10.4
Werkzeug~=2.3.6
zipp==3.10.0
pandas~=2.0.3
qrcode~=7.4.2
redis~=4.6.0
rq~=1.15.1
SQLAlchemy~=1.4.22
pydantic~=2.0.3
alembic~=1.11.1
pip~=23.2
setuptools~=52.0.0
python-dotenv~=1.0.0